made vaccinations dir; updated imports; separated Add vax dialog to another file
