+++ src/components/medicals/ViewMedical.js
@@ -7,7 +7,7 @@ import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import { MedicationsList } from "./medications/MedicationsList";
-import { VaccinationsList } from "./VaccinationsList";
+import { VaccinationsList } from "./vaccinations/VaccinationsList";
 
 export const ViewMedical = ({pet, syncPets}) => {
